deleted:    sklift/tests/test_plot_qini_curve.py 	deleted:    sklift/tests/test_plot_uplift_curve.py 	new file:   sklift/tests/test_viz.py
